requirements.txt
gradio
pytesseract
pillow
python-docx
opencv-python
